Extracted RunnerExtension
