+++ lib/PhpBench.php
@@ -18,6 +18,7 @@ use PhpBench\DependencyInjection\Container;
 use PhpBench\Exception\ConfigurationPreProcessingError;
 use PhpBench\Extension\CoreExtension;
 use PhpBench\Extension\ExpressionExtension;
+use PhpBench\Extension\RunnerExtension;
 use PhpBench\Extensions\XDebug\XDebugExtension;
 use PhpBench\Json\JsonDecoder;
 use Seld\JsonLint\JsonParser;
@@ -61,6 +62,7 @@ class PhpBench
 
         $extensions = $config['extensions'];
         $extensions[] = CoreExtension::class;
+        $extensions[] = RunnerExtension::class;
         $extensions[] = ExpressionExtension::class;
 
         if (extension_loaded('xdebug')) {
